Massive Push - Majority Back-End (See details)
+ Encapsulated Database Access
++ DB Connection implemented into a classs, constructor creates the PDO
connection
++ Create, Read, Update and Delete classes created that extend the DB
Connection Class, these classes consist of now more then two functions
inorder to achieve their purpose.
++  Create, Read, Update and Delete classes each have an object class
that they extend too, these object classes define the paramenters and
set the query in which to run
+ Added the removal of a Wine product from the administration page as
we'll as asking for confirmation
+ Added new attribute to the database to indicate the Wine Taste, for
red and white wines.
+ Added functionality to include the inidicator when adding a new wine.
+ Added functionality to amend the indicator when updating existing
wines.
+ Added a filter for the Wine Task on the wine product page that
filters.
+ Enforced Campaign display, which now only presents on the wine page
Wines that are on offer within the range dates.
+ Added a return to the customer home page from the Administration Pages
+ Restyled mobile navigation in the Administration pages
+  Added to the creation of a new wine, the date added attribute to the
database this is to be used within the dynamic RSS Feed, and now makes
this more accurate

++++ UPDATED DATABASE FILE ++++
Please use the latest version